[DOP-8140] Split MongoDB.Dialect and .*Options to separated submodules
