--- docs/connection/db_connection/mongodb.rst
@@ -1,26 +1,0 @@
-.. _mongo:
-
-MongoDB connection
-=====================
-
-.. currentmodule:: onetl.connection.db_connection.mongodb
-
-.. autosummary::
-
-    MongoDB
-    MongoDB.ReadOptions
-    MongoDB.WriteOptions
-    MongoDB.PipelineOptions
-
-.. autoclass:: MongoDB
-    :members: get_packages, check, pipeline
-
-.. currentmodule:: onetl.connection.db_connection.mongodb.MongoDB
-
-.. autopydantic_model:: ReadOptions
-
-.. autopydantic_model:: WriteOptions
-    :members: mode
-    :member-order: bysource
-
-.. autopydantic_model:: PipelineOptions
