+++ docs/connection/db_connection/mongodb/connection.rst
@@ -1,0 +1,10 @@
+.. _mongodb-connection:
+
+MongoDB Connection
+==================
+
+.. currentmodule:: onetl.connection.db_connection.mongodb.connection
+
+.. autoclass:: MongoDB
+    :members: get_packages, check
+    :member-order: bysource
