+++ docs/connection/db_connection/mongodb/index.rst
@@ -1,0 +1,17 @@
+.. _mongodb:
+
+MongoDB
+=======
+
+.. toctree::
+    :maxdepth: 1
+    :caption: Connection
+
+    connection
+
+.. toctree::
+    :maxdepth: 1
+    :caption: Operations
+
+    read
+    write
