+++ docs/connection/db_connection/mongodb/write.rst
@@ -1,0 +1,13 @@
+.. _mongodb-write:
+
+Writing to MongoDB
+==================
+
+For writing data to MongoDB, use :obj:`DBWriter <onetl.db.db_writer.db_writer.DBWriter>` with options below.
+
+.. currentmodule:: onetl.connection.db_connection.mongodb.options
+
+.. autopydantic_model:: MongoDBWriteOptions
+    :member-order: bysource
+    :model-show-field-summary: false
+    :field-show-constraints: false
