+++ setup.cfg
@@ -307,9 +307,6 @@ per-file-ignores =
     *connection.py:
 # WPS437 Found protected attribute usage: spark._sc._gateway
         WPS437,
-    onetl/connection/db_connection/mongodb.py:
-# WPS437 Found protected attribute usage: self.Dialect._
-        WPS437,
     onetl/connection/db_connection/jdbc_mixin/connection.py:
 # too few type annotations
         TAE001,
